requirements/requirements.top.txt
urllib3
requests
boto3
botocore
pip
numpy
pandas
setuptools
wheel
pyyaml
six
python-dateutil
typing-extensions
s3fs
cryptography
certifi
charset-normalizer
idna
packaging
pyjwt
matplotlib
scipy
tensorflow
scikit-learn
pytorch
beautifulsoup4
pillow
fastapi
django
flask
